Consolidate mode helpers under android_ms11
