+++ android_ms11/modes/whisper_mode.py
@@ -1,3 +1,14 @@
+def start_buff_by_tell(character_name: str) -> None:
+    """Simulate responding to buff requests sent via tell."""
+
+    print(f"📨 Buff-by-Tell Mode Activated for {character_name}...")
+    print("🧙 Listening for buff requests and queueing macros...")
+    # Simulate test output
+    print("📨 Simulation: Responded to 2 buff requests.")
+
+
 def run(config, session=None):
     """Main entry point for this mode."""
-    pass
+
+    character = config.get("character_name", "Unknown") if config else "Unknown"
+    start_buff_by_tell(character)
