--- src/mode_afk_combat.py
@@ -1,14 +1,0 @@
-
-"""AFK combat automation helpers.
-
-The real AFK combat system is not implemented yet. This stub simply
-prints messages that mimic basic target rotation and whisper monitoring.
-"""
-
-def start_afk_combat(character_name: str) -> None:
-    """Simulate passive combat actions for a character."""
-
-    print(f"🛡️ AFK Combat Mode Activated for {character_name}...")
-    print("⚔️ Rotating through targets and listening for whispers...")
-    # Simulate test output
-    print("⚔️ Simulation: Defeated 5 mobs at waypoint.")
