--- src/mode_buff_by_tell.py
@@ -1,14 +1,0 @@
-
-"""Utilities for buff-by-tell mode.
-
-The full buff-by-tell system is still under construction. This stub
-demonstrates the expected console output when a request is handled.
-"""
-
-def start_buff_by_tell(character_name: str) -> None:
-    """Simulate responding to buff requests sent via tell."""
-
-    print(f"📨 Buff-by-Tell Mode Activated for {character_name}...")
-    print("🧙 Listening for buff requests and queueing macros...")
-    # Simulate test output
-    print("📨 Simulation: Responded to 2 buff requests.")
