--- src/mode_crafting.py
@@ -1,14 +1,0 @@
-
-"""Automation routines for crafting mode.
-
-The crafting system is largely unimplemented. For now the helper simply
-prints messages describing the high level actions that would occur.
-"""
-
-def start_crafting(character_name: str) -> None:
-    """Simulate a short crafting session for the given character."""
-
-    print(f"🛠️ Crafting Mode Activated for {character_name}...")
-    print("🔧 Surveying resources and executing crafting macros...")
-    # Simulate test output
-    print("🔧 Simulation: Crafted 2 Mineral Survey Devices.")
