--- src/mode_medic.py
@@ -1,15 +1,0 @@
-
-"""Medic mode automation helpers.
-
-This module currently provides a very small stub that simulates basic
-healing behaviour. The real automation logic for loading macros and
-monitoring group status has not been implemented yet.
-"""
-
-def start_medic(character_name: str) -> None:
-    """Simulate healing and buffing nearby allies."""
-
-    print(f"🩺 Medic Mode Activated for {character_name}...")
-    print("🧪 Monitoring group status and applying basic heals...")
-    # Simulate test output
-    print("🧪 Simulation: Buffed 3 nearby players.")
