--- src/mode_questing.py
@@ -1,8 +1,0 @@
-"""Questing mode utilities."""
-
-def start_questing(character_name: str):
-    """Placeholder to simulate questing steps."""
-    print(f"Questing for {character_name}...")
-    # In a real implementation this would run game automation
-    steps = ["start", "finish"]
-    return steps
