+++ tests/test_mode_stubs.py
@@ -3,10 +3,10 @@ import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from src.mode_medic import start_medic
-from src.mode_afk_combat import start_afk_combat
-from src.mode_crafting import start_crafting
-from src.mode_buff_by_tell import start_buff_by_tell
+from android_ms11.modes.medic_mode import start_medic
+from android_ms11.modes.combat_assist_mode import start_afk_combat
+from android_ms11.modes.crafting_mode import start_crafting
+from android_ms11.modes.whisper_mode import start_buff_by_tell
 
 
 def test_start_medic_outputs_stub(capsys):
